squareImg() @Transactional 작동하도록 수정한다
